- Увеличена выручка с ограбления магазинов в 2 раза - Починил HK-416 - Стоимость лечения для опытных игроков (Которые уже отыграли какое-то кол-во часов на сервере) была снижена до $300 (Без мед. страховки) - Обновлена одежда (Не добавлена) - Обновлены позиции фургонов войны за грузы и добавлены новые - Обновлен опыт и зарплаты на работе автобуса (В большую сторону) - Обновлено получение опыта на работах садовник и разнорабочий (В большую сторону) - Обновлена зарплата и опыт на работе фотографа (В большую сторону) - В бизнесе теперь можно ставить более точные скидки (Например 133%) - Убрана возможность EMS эвакуировать на штраф стоянку - Фикс бага с катастрофами (Не выбирались рандомные) - Куча мелких фиксов и не очень мелких, но вам знать не надо о них)

+++ src_client_package/src/property/business.js
@@ -112,7 +112,7 @@ business.getSale = function(price) {
 };
 
 business.setPrice = function(id, price) {
-    Container.Data.Set(enums.offsets.business + id, 'price_product', methods.parseInt(price));
+    Container.Data.Set(enums.offsets.business + id, 'price_product', methods.parseFloat(price));
 };
 
 business.setName = function(id, name) {
